+++ document-service/src/main/java/com/buaa01/illumineer_backend/service/paper/PaperSearchService.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 public interface PaperSearchService {
 
-    CustomResponse getPaperByPid(Integer pid);
+    CustomResponse getPaperByPid(Long pid);
 
     CustomResponse getPaperByStats(Integer stats, Integer size, Integer offset, Integer sortType, Integer order);
